test/web-platform-tests/to-upstream/html/semantics/forms/the-textarea-element/setSelectionRange.html

<!DOCTYPE HTML>
<title>textarea element setSelectionRange() functionality</title>
<link rel="author" title="yaycmyk" href="mailto:evan@yaycmyk.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#textFieldSelection">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  "use strict";

  test(() => {
    const element = document.createElement("textarea");

    element.value = "foo";
    element.setSelectionRange(0, 2);

    assert_equals(element.selectionStart, 0);
    assert_equals(element.selectionEnd, 2);
    assert_equals(element.selectionDirection, "none");

  }, "textarea setSelectionRange(start, end) should select a specific character range");

  async_test(t => {
    const element = document.createElement("textarea");

    element.value = "foo";

    element.addEventListener("select", () => {
      t.step(() => {
        assert_true(true);
        t.done();
      });
    });

    element.setSelectionRange(0, 2);

  }, "textarea setSelectionRange() should emit a select event");
</script>
